requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1